requirements.txt
fastapi
uvicorn
requests
psycopg2-binary
PyGithub
dotenv